+++ CHANGELOG.rst
@@ -4,6 +4,24 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
+1.11.0 (2021-05-20)
+-------------------
+
+**Added**
+
+- Metabolomics import format
+- "QBiC" format has been renamed to openBIS-based format
+- "Standard" format has been renamed to "Standard QBiC format"
+
+**Fixed**
+
+- Help/Examples symbols for import formats are now places directly next to the respective options
+- When importing multiple different formats, the project and experimental design are now correctly reset
+
+**Dependencies**
+
+**Deprecated**
+
 
 1.10.1 (2021-02-02)
 -------------------